server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
numpy>=1.24,<2
pillow>=10,<12
onnx>=1.15
onnxruntime>=1.17
torch==2.2.0
torchvision==0.17.0
pyclipper>=1.3
shapely>=2.0
lxml>=4.9
requests>=2.28,<3
